Update German translations
The prisoner is indeed referred to as "Häftling", fixed grammar, their sarcophagus is referenced to as "Kammer" (chamber).

I renamed the entrances as the given words are grammatically wrong unless this was in the ship log as "given name" for them.

While we are at it, I really dislike the German name given for the dam. I could not find ingame references to it so I am skipping this one.